requirements.txt
# Core dependencies
pandas>=1.3.0
pyyaml>=5.4.0

# Web dashboard (optional)
flask>=2.0.0
plotly>=5.0.0

# Calendar integration (optional)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0